moved interfaces to same file
