--- API/Services/ILibraryService.cs
@@ -1,10 +1,0 @@
-using API.DTOs;
-
-namespace API.Services;
-public interface ILibraryService : IServiceBase
-{
-    List<LibraryDTO> GetUserLibrariesById();
-    LibraryDTO GetUserLibraryById();
-    List<UploadDTO> GetLibraryUploadsById();
-    UploadDTO GetLibraryUploadById();
-}
