--- API/Services/IUploadService.cs
@@ -1,5 +1,0 @@
-namespace API.Services;
-public interface IUploadService : IServiceBase
-{
-    string StoreUpload();
-}
