--- API/Services/IUserCredentialsService.cs
@@ -1,12 +1,0 @@
-using API.DTOs;
-
-namespace API.Services;
-public interface IUserCredentialsService : IServiceBase
-{
-    bool CheckIfUserExistsByEmail();
-    void StoreNewUserCredentials();
-    bool ValidateUserCredentials();
-    UserDTO GetUserDataByEmail();
-    void StoreUserSession();
-    UserDTO GetUserDataBySession();
-}
