+++ API/Services/LibraryService.cs
@@ -1,6 +1,13 @@
 using API.DTOs;
 
 namespace API.Services;
+public interface ILibraryService : IServiceBase
+{
+    List<LibraryDTO> GetUserLibrariesById();
+    LibraryDTO GetUserLibraryById();
+    List<UploadDTO> GetLibraryUploadsById();
+    UploadDTO GetLibraryUploadById();
+}
 public class LibraryService : ILibraryService
 {
     /// <summary>
