+++ API/Services/UserCredentialsService.cs
@@ -1,6 +1,15 @@
 using API.DTOs;
 
 namespace API.Services;
+public interface IUserCredentialsService : IServiceBase
+{
+    bool CheckIfUserExistsByEmail();
+    void StoreNewUserCredentials();
+    bool ValidateUserCredentials();
+    UserDTO GetUserDataByEmail();
+    void StoreUserSession();
+    UserDTO GetUserDataBySession();
+}
 public class UserCredentialsService : IUserCredentialsService
 {
     /// <summary>
